requirements.txt
fastapi

uvicorn[standard]

sqlalchemy==2.0.*

psycopg[binary]

pydantic

python-jose[cryptography]

passlib[bcrypt]

python-dotenv

celery

redis